User-modifiable gain. At what level do we want to handle this?
I feel like settings such as gain, duration, and spacing may need to be at a different level. Are they specific to a pitch map (and set of pyo etc arguments), or individual sonified objects? Are they general astronify settings? As we start making these user-modifiable this is something we'll have to move, and I'd appreciate input as someone who just picked this up.